+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Versionning adheres to [Semantic Versioning] (https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 ### Added
+- add basic support to fetch metadata from google books api
 - generate an help file in markdown format in addition to the manpage
 
 ## [0.1.0] - 2019-05-10